:recycle: REFACTOR: 글작성 페이지 헤더 분리
