+++ src/pages/Write.jsx
@@ -1,6 +1,6 @@
 import { useState, useRef, useCallback, useEffect } from "react";
 import styled, { css } from "styled-components";
-import { StContainer, StHeader, StSection } from "../UI/common";
+import { useParams } from "react-router-dom";
 import { useMutation } from "@tanstack/react-query";
 import { postsApi } from "../apis/axios";
 
@@ -9,11 +9,10 @@ import HashTagInput from "../components/common/HashTagInput";
 import NavigateBtn from "../components/common/NavigateBtn";
 import TextEditor from "../components/common/TextEditor";
 import WeatherPicker from "../components/write/WeatherPicker";
-import WritePageTutorialModal from "../components/write/WritePageTutorialModal";
-import { useNavigate, useParams } from "react-router-dom";
+import WritePageSubmitHeader from "../components/write/WritePageSubmitHeader";
+import { StContainer, StHeader, StSection } from "../UI/common";
 import { imgUrlConvertBlob } from "../utils/imgUrlConvertBlob";
 import useDispatchHook from "../hooks/useDispatchHook";
-import { BsQuestionLg } from "react-icons/bs";
 
 const Write = () => {
   const [canvas, setCanvas] = useState("");
@@ -61,7 +60,6 @@ const Write = () => {
 
     formData.get("title");
     formData.get("createdAt");
-
     formData.append("image", blob, "img.file");
     formData.append("content", contents);
     formData.append("weather", weather || "sun");
@@ -76,32 +74,6 @@ const Write = () => {
     else setIsDrawingEnd(!isDrawingEnd);
   }, []);
 
-  const defaultHeader = () => {
-    return (
-      <>
-        <NavigateBtn prev link={`/list/${diaryId}`} />
-        <h3>LOGO</h3>
-        <span onClick={nextSectionHeaderHandler}>다음</span>
-      </>
-    );
-  };
-
-  const drawingEndHeader = () => {
-    return (
-      <>
-        <span onClick={() => setIsDrawingEnd(!isDrawingEnd)}>뒤로가기</span>
-        <StDrawindEndHeaderOptionBox>
-          <WritePageTutorialModal>
-            <StQuestionIcon />
-          </WritePageTutorialModal>
-
-          <StWriteFormSubmitBtn type="submit" form="writeForm">
-            완성
-          </StWriteFormSubmitBtn>
-        </StDrawindEndHeaderOptionBox>
-      </>
-    );
-  };
   const preventTabKey = (event) => {
     event.key === "Tab" && event.preventDefault();
   };
@@ -112,11 +84,31 @@ const Write = () => {
     return () => document.removeEventListener("keydown", preventTabKey);
   }, []);
 
+  const StDiv = styled.div`
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 1rem;
+  `;
+
   return (
     <>
       <StContainer>
         <StHeader flex justify="space-between" aline="center">
-          {isDrawingEnd ? drawingEndHeader() : defaultHeader()}
+          {isDrawingEnd ? (
+            <WritePageSubmitHeader
+              setIsDrawingEnd={setIsDrawingEnd}
+              isDrawingEnd={isDrawingEnd}
+            />
+          ) : (
+            <>
+              <StDiv>
+                <NavigateBtn prev link={`/list/${diaryId}`} />
+                <h4>다이어리 생성</h4>
+              </StDiv>
+              <span onClick={nextSectionHeaderHandler}>다음</span>
+            </>
+          )}
         </StHeader>
         <StSlideWrapper isDrawingEnd={isDrawingEnd}>
           <StTextSection>
@@ -231,23 +223,4 @@ const StSlideWrapper = styled.div`
     css`
       transform: translateX(-50%);
     `}
-`;
-
-const StWriteFormSubmitBtn = styled.button`
-  color: #3cc7a6;
-  cursor: pointer;
-  border: none;
-  background-color: transparent;
-  font-size: 1.6rem;
-`;
-
-const StDrawindEndHeaderOptionBox = styled.div`
-  display: flex;
-  gap: 1rem;
-`;
-
-const StQuestionIcon = styled(BsQuestionLg)`
-  font-size: 2.4rem;
-  color: var(--grayscale_5);
-  cursor: pointer;
 `;
